sensor-node/requirements.txt
mpu6050-raspberrypi>=1.1
smbus2>=0.4.1
RPi.bme280>=0.2.4
adafruit-circuitpython-ccs811>=1.3
paho-mqtt>=1.6.1
pyyaml>=6.0
